pyproject.toml
[project]
name = "django-estudo"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=70.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
    "drf-spectacular>=0.27.2",
    "python-dotenv>=1.0.1",
    "netifaces>=0.11.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
